Refactor session end receiver
