+++ dntu_focus/android/app/src/main/AndroidManifest.xml
@@ -59,7 +59,7 @@
 
         <!-- Session End Receiver -->
         <receiver
-            android:name=".MainActivity$sessionEndReceiver"
+            android:name=".SessionEndReceiver"
             android:exported="false">
             <intent-filter>
                 <action android:name="com.example.moji_todo.SESSION_END" />
